Rehome account hook helpers
